Add files via upload
Here is a module I created that tracks the persistent H₀ and H₁ homologies of simplexes generated by the SIFT features of a 2D image.

The three inputs of this function are your image, (steps) the steps taken -or how many columns of data you have-, and (stepsize) the step size. If two keypoints have intersecting radii, then there will be a one simplex between them. If three keypoints have intersecting radii, then there will be a two simplex between them. The function creates a table of the radius, the H₀ Betti number, and the H₁ Betti number.